Move ui_component_value_editor
to common.c.
Improvements can continue from there.
