+++ src/menu/ui_components.h
@@ -278,4 +278,15 @@ void ui_components_tabs_draw(const char **text, int count, int selected, float w
  */
 void ui_components_tabs_common_draw(int selected);
 
+/**
+ * @brief Draw a value editor component.
+ * 
+ * @param header_text Array of header text for the values.
+ * @param value_text Array of value text to be displayed.
+ * @param count Number of values.
+ * @param selected Index of the selected value.
+ * @param width Width of each value box.
+ */
+void ui_component_value_editor(const char **header_text, const char **value_text, int count, int selected, float width );
+
 #endif /* UI_COMPONENTS_H__ */
